app/requirements.txt
google-genai==1.0.0
google-auth==2.25.2
google-api-python-client==2.110.0